feat: Add support for adding layers to the canvas
This commit adds the functionality to insert new layers to the canvas. It introduces a new `insertLayer` mutation in the `Canvas` component, which allows users to add layers of different types (Ellipse, Rectangle, Note, Text) at a specified position. The maximum number of layers is limited to 100. The `lastUsedColor` state is also updated when inserting a new layer. This feature enhances the canvas by enabling users to create and organize different elements on the board.

Note: No newline at end of file