feat: Add bring to front and send to back functionality in SelectionTools component
This commit adds the functionality to bring selected layers to the front and send them to the back in the SelectionTools component. It introduces two new buttons, "Bring to front" and "Send to back", which allow users to change the stacking order of the selected layers on the canvas. This enhancement provides users with more control over the layer arrangement and improves the overall usability of the application.